+++ spec/dummy/app/javascript/Globals.d.ts
@@ -1,3 +1,9 @@
-declare module "*.module.css";
-declare module "*.module.scss";
+declare module "*.module.css" {
+  const classes: { [key: string]: string };
+  export = classes;
+}
+declare module "*.module.scss" {
+  const classes: { [key: string]: string };
+  export = classes;
+}
 declare module "*.svg";